# api.yml
openapi: 3.0.3
info:
  title: API
  description: |
    API documentation including:
    - **Auth** – login, register, refresh (public); returns JWT for protected routes.
    - **User** – profile, badges, task history, streak, resume, profile pic (JWT required).
    - **Tasks** – get assigned tasks with **user_status** (completed, viewing, rejected, not_started); submit proof (JWT required).
    - **Leaderboard** – pan-india, state, college; each with optional period (all, weekly, monthly) or dedicated /weekly and /monthly routes; entry has name, rank, xp, profile_image, id, state, college.
    - **States** – list all states (public).
    - **Colleges** – list colleges by state (public).
  version: 1.0.0
servers:
  - url: /api
    description: API base path
security:
  - BearerAuth: []
paths:
...
  /leaderboard/pan-india:
    get:
      summary: Get pan-India leaderboard
      description: |
        Get the pan-India leaderboard. Returns name, rank, xp, profile image, id, state, and college for each entry.
        Frontend can filter by state or college using state_id and college_id. No authentication required.
      operationId: getPanIndiaLeaderboard
      tags:
        - leaderboard
      security: []
      parameters:
        - name: page
          in: query
          required: false
          description: Page number (default 1)
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          required: false
          description: Items per page (default 100, max 1000)
          schema:
            type: integer
            default: 100
        - name: period
          in: query
          required: false
          description: Time period – all, weekly, monthly
          schema:
            type: string
            enum: [all, weekly, monthly]
            default: all
      responses:
        '200':
          description: Pan-India leaderboard entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardResponse'
              example:
                entries:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    name: "Alice"
                    rank: 1
                    xp: 1500
                    profile_image: "https://bucket.s3.amazonaws.com/profile/alice.jpg"
                    state_id: "state-uuid-1"
                    state_name: "Maharashtra"
                    college_id: "college-uuid-1"
                    college_name: "Example College"
                  - id: "660e8400-e29b-41d4-a716-446655440001"
                    name: "Bob"
                    rank: 2
                    xp: 1200
                    profile_image: ""
                    state_id: "state-uuid-2"
                    state_name: "Karnataka"
                    college_id: "college-uuid-2"
                    college_name: "Another College"
                type: pan-india
                page: 1
                page_size: 100
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
...
  /leaderboard/state:
    get:
      summary: Get state leaderboard
      description: Get leaderboard for a specific state. Same entry shape (name, rank, xp, profile_image, id, state, college). No authentication required.
      operationId: getStateLeaderboard
      tags:
        - leaderboard
      security: []
      parameters:
        - name: state_id
          in: query
          required: true
          description: State ID (UUID)
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            default: 100
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [all, weekly, monthly]
            default: all
      responses:
        '200':
          description: State leaderboard entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardResponse'
        '400':
          description: Bad request – state_id required
        '500':
          description: Internal server error
...
  /leaderboard/college:
    get:
      summary: Get college leaderboard
      description: Get leaderboard for a specific college. Same entry shape (name, rank, xp, profile_image, id, state, college). No authentication required.
      operationId: getCollegeLeaderboard
      tags:
        - leaderboard
      security: []
      parameters:
        - name: college_id
          in: query
          required: true
          description: College ID (UUID)
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            default: 100
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [all, weekly, monthly]
            default: all
      responses:
        '200':
          description: College leaderboard entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardResponse'
        '400':
          description: Bad request – college_id required
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    LeaderboardResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/LeaderboardEntry'
        type:
          type: string
          enum: [pan-india, state, college]
        scope_id:
          type: string
          description: state_id or college_id (empty for pan-india)
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
          description: Optional total count
    LeaderboardEntry:
      type: object
      description: One leaderboard row – name, rank, xp, profile image, id, state, college
      properties:
        id:
          type: string
          format: uuid
          description: User ID
        name:
          type: string
        rank:
          type: integer
        xp:
          type: integer
        profile_image:
          type: string
          description: Profile/avatar URL (optional)
        level:
          type: integer
        state_id:
          type: string
          format: uuid
        state_name:
          type: string
        college_id:
          type: string
          format: uuid
        college_name:
          type: string
...